requirements.txt
diffusers==0.2.4
transformers
scipy
ftfy
ipywidgets>=7,<8
torch
opencv-python
pillow
google